feat: Update payroll details in EditPayrolls view and controller
This commit updates the EditPayrolls view and controller to include the ability to update payroll details. It adds a new route 'UpdatePayroll' in web.php and a new method 'UpdatePayroll' in PayrollsController.php. The Payroll_details model is also updated with a new static method 'UpdatePayroll' to handle the update request. The changes in the view include adding a button to save the updated payroll details and adding JavaScript code to handle the AJAX request. The changes in the controller include handling the update request and returning a JSON response. This feature allows users to update the commission and days worked for a payroll.

Recent user commits:
- [ADD] inactive employee display issue
- [ADD] Work progress
- [ADD] getDashboard method to only require dt_end parameter
- [ADD] Mejoras a la UI de Modulos financieros
- [feat] Add Consolidado route and update lyt_aside.blade.php
- [feat] add Consolidado route and update lyt_aside.blade.php
- [ADD] route for editing payrolls by employee ID
- [FEAT] inicio del modulo de planilla
- [FIX]
- Fix
- [FIX] typo nombre ruta
- [Feat] CRUD de gastos Operativos
- [ADD] Filtrado de fecha Dashboard
- [ADD] Nuevos permisos para promotor
- [ADD] Permisos de ediciion de creditos
- [ADD] Proder Editar un credito